Initialize struct to resolve caution of Rule 9.1.
The variable 'temp' may not be set when passed to 'sensor_channel_get',
if the function fails. In this case, it the code would be safe if and
only if it does not access 'temp' afterwards, but to avoid undefined
behavior the simplest solution is to initalize the struct at the
declaration.

Signed-off-by: Nicola Vetrini <nicola.vetrini@bugseng.com>